Hub/requirements.txt
fastapi
uvicorn
python-multipart
pandas
pyarrow
psycopg2-binary
matplotlib
